Update version number location in deploy.sh
The script `deploy.sh` has been modified to export the version number earlier in the sequence of operations. By doing this, the version number is available for use sooner, and unnecessary repetition of the operation later in the script is prevented. The version number is now determined just after command line arguments have been parsed.