Add OpenCDXCurrentUser to ClassifyProcessorService and update tests
In `OpenCDXClassifyProcessorServiceImpl.java`, `OpenCDXCurrentUser` was added as a dependency to get user access tokens for media downloading. Changes were made to the constructor and download function to use this new dependency. Unit tests were updated in `OpenCDXClassificationServiceImplTest.java` to reflect these changes and a new authorization parameter was added in `OpenCDXMediaUpDownClient.java` for file upload and download methods.