Add auto-confirm option to build Docker images
Added an 'auto_confirm_all' option to the 'build_docker' method in the deployment script. This optional parameter, when set to true, bypasses the component selection process and directly proceeds in building images for all available components. This allows users to streamline the deployment process when building all components is desired.