Refactor AppConfig and introduce OpenCDXClassificationMessageHandler
AppConfig.java has been refactored to replace the format bean and instead instantiate the new OpenCDXClassificationMessageHandler. This handler has been created to manage incoming classification messages and the corresponding service. This refactoring also involved the deletion of AppConfigTest.java. The changes are part of a larger effort to streamline and improve our message handling systems.