Refactor UserProfile id field and improve profile creation logic
Refactoring involved marking "id" field in UserProfile as optional and updating related usage in service implementation. Added a test method for CreateUserProfileRequest in IamTest and removed a redundant test from OpenCDXIAMProfileServiceImplTest. Also, the OpenCDXProfileModel constructor was modified to update id from UserProfile, and the update and model creation logic in service implementation was refined.