Add 'Shipping' microservice to OpenCDX
A new 'Shipping' microservice was added to the OpenCDX project, together with basic configuration and structure. The microservice named as 'opencdx-shipping' contains model, service, controller, repository and config. This microservice was indexed in the build.gradle file. It was also included as a component of the deployment script deploy.sh.