Add multiple new ShippingVendor classes
Added four ShippingVendor classes (UpsShippingVendor, UspsShippingVendor, FedexShippingVendor and DoorDashShippingVendor) and modified the getShippingVendors and shipPackage methods in the OpenCDXShippingVendorServiceImpl. Also, the OpenCDXShippingVendor interface and OpenCDXShippingModel class are newly added and Shipping protocol buffer message is revised. This update allows the system to correctly select the appropriate shipping vendor to obtain shipping packages options and ship packages.