Add LIDR ID field to OpenCDX device and test case models
Implemented a unique identifier for the Laboratory Interoperability Data Repository (LIDR) ID in the OpenCDX device and test case models. This includes changes in the associated protobuf files, model classes, and unit tests to support and validate the new LIDR ID field. Moreover, it adjusts the numbering of other fields in the protobuf files accordingly.