Add messageType in OpenCDXMessageModel and change HTTP methods for /messages endpoint
The messageType attribute is now being saved in the OpenCDXMessageModel, enhancing details about the communication messages. The /messages endpoint changes methods from HTTP GET to POST in both Controller and Test. The change in the HTTP methods allows for better compliance with REST guidelines while ensuring a more secure way of sending requests with larger bodies.