Add DosageForm enum and update Medication proto properties
Introduced a new DosageForm enum in the 'health_medication.proto' file which contains different dosage forms of medication. Also expanded the Medication message to include new properties such as 'dosage_form' and 'quantity', and updated the indexing of existing properties for consistency. Additionally, included MedicationService for handling medication-related tasks, and created several request and response messages for handling the operations.