Implement Medication management services in OpenCDX
This update includes multiple new additions to the OpenCDX system for managing medications. A new Medication model was created in the OpenCDXMedicationModel class. Alongside this, a service was introduced (OpenCDXMedicationService) and its implementation (OpenCDXMedicationServiceImpl) to handle medication-related operations. Furthermore, a controller (OpenCDXMedicationGrpcController) was set up to handle gRPC calls. This also includes changes in protobuf definitions with the addition of an RESTful API endpoint.