Improve error handling in getMedicationsByBrandName method
Updated the getMedicationsByBrandName method in the OpenCDXApiFDAImpl class to enhance error handling. The program now catches FeignExceptions and logs a warning when there is an issue fetching drug data based on the brand name, then returns an empty list instead of a null value. This improvement increases the reliability of data retrieval processes and the robustness of the system.