Integrate OpenFDA API Key in OpenCDX system.
The OpenCDX system now supports the use of an OpenFDA API Key in its services. This update was implemented by modifying the OpenCDXApiFDAImpl and OpenCDXOpenFDAClient classes and adding the key to application.yml file. The key is utilized in the getDrugs and getLabel methods, enhancing the security of data retrieval from the OpenFDA system.