Integrate KnowledgeService as class variable in `OpenCDXAnalysisEngineImpl`
This commit adds KnowledgeService as a class variable in `OpenCDXAnalysisEngineImpl`. This tweak not only makes the class easier to read but can also enhance performance by preventing regular instantiation of KnowledgeService in the method body each time it's called. Instead, a single instance is used throughout the class life cycle, improving efficiency.