Update JMeter configuration parameters
Several changes have been made in the JMeter properties and configuration. The rampUp value and lCount value have been changed from 0 to 1 in the smoke.properties file. In addition, the ThreadGroup scheduler has been set to 'false', and the default number of threads and ramp time have been modified in the OpenCDX.jmx file.