Refactor gRPC imports across multiple applications
This commit addresses the refactoring of gRPC imports across several applications within the opencdx-commons project. Changes are focused on adjusting to a new structure in gRPC services and types, including those related to IAM, health profile, media, and audit, among others. This improves the clarity of the file structure and makes it easier to maintain the codebase.