Add Node.js check to deploy.sh and update README
Added a check for Node.js in the deployment script and updated the software list in the README file. The README now includes OpenSSL and Node.js in its software list. Also, the deployment script now checks that the correct version of Node.js is installed during the deployment process.