Refactored medical conditions service for specific ID operations
In the medical conditions base code, new messages have been added to the protobuf that allows operations based on specific diagnosis ID. Consequentially, the relevant interfaces, implementation, and controllers have been updated to include these new methods. Moreover, an "update" function was added in the model to simplify the process of updating a diagnosis.