Refactor protobuf types and remove CircumstancePriority enum
Changed several fields in the opencdx_data.proto from specific data types to the more generic Measure type. This change simplifies the code, reducing the number of fields associated with keeping track of time durations and frequencies. The actions that were previously detailed, such as period_start, event_frequency, are now more generic and can be used in a wider range of applications. Also, the CircumstancePriority enum in opencdx_enum.proto was removed, as it has now been replaced with a simpler string type in the opencdx_data.proto.